And there was much renaming
