--- server/src/main/java/org/apache/druid/catalog/MeasureColumnDefn.java
@@ -1,48 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.druid.catalog;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-/**
- * Catalog definition of a measure (metric) column.
- */
-public class MeasureColumnDefn extends DatasourceColumnDefn
-{
-  private final String aggregateFn;
-
-  @JsonCreator
-  public MeasureColumnDefn(
-      @JsonProperty("name") String name,
-      @JsonProperty("sqlType") String sqlType,
-      @JsonProperty("aggregateFn") String aggregateFn
-  )
-  {
-    super(name, sqlType);
-    this.aggregateFn = aggregateFn;
-  }
-
-  @JsonProperty("aggregateFn")
-  public String aggregateFn()
-  {
-    return aggregateFn;
-  }
-}
