+++ server/src/main/java/org/apache/druid/metadata/catalog/CatalogManager.java
@@ -19,13 +19,23 @@
 
 package org.apache.druid.metadata.catalog;
 
+import org.apache.druid.catalog.TableId;
+import org.apache.druid.catalog.TableMetadata;
+import org.apache.druid.catalog.TableSpec;
 import org.apache.druid.java.util.common.ISE;
+
+import javax.annotation.Nullable;
+
+import java.util.List;
 
 /**
  * Manages catalog data. Used in Coordinator, which will be in either
  * an leader or standby state. The Coordinator calls the {@link #start()}
  * method when it becomes the leader, and calls {@link #stop()} when
  * it loses leadership, or shuts down.
+ *
+ * Performs detailed CRUD operations on the catalog tables table.
+ * Higher-level operations appear elsewhere.
  */
 public interface CatalogManager
 {
@@ -92,9 +102,84 @@ public interface CatalogManager
     }
   }
 
+  interface Listener
+  {
+    void added(TableMetadata table);
+    void updated(TableMetadata table);
+    void deleted(TableId id);
+  }
+
   void start();
 
-  void stop();
 
-  TableDefnManager tables();
+  void register(Listener listener);
+  void createTableDefnTable();
+
+  /**
+   * Create a table entry.
+   *
+   * @return the version of the newly created table. Call
+   * {@link TableMetadata#asUpdate(long)} if you want a new
+   * {@link TableMetadata} with the new version.
+   * @throws {@link DuplicateKeyException} if the row is a duplicate
+   * (schema, name) pair. This generally indicates a code error,
+   * or since our code is perfect, a race condition or a DB
+   * update outside of Druid. In any event, the error is not
+   * retryable: the user should pick another name, or update the
+   * existing table
+   */
+  long create(TableMetadata table) throws DuplicateKeyException;
+
+  /**
+   * Update a table definition, but only if the database entry is at
+   * the given {@code oldVersion}.
+   */
+  long updateDefn(TableId id, TableSpec defn, long oldVersion) throws OutOfDateException;
+
+  /**
+   * Update a table definition, overwriting any current content.
+   * This is a potential race conditions if this is a partial update
+   * because of the possibility of another user doing an update since the
+   * read. Fine when the goal is to replace the entire definition.
+   */
+  long updateDefn(TableId id, TableSpec defn) throws NotFoundException;
+
+  /**
+   * Move the table to the deleting state. No version check: fine
+   * if the table is already in the deleting state. Does nothing if the
+   * table does not exist.
+   *
+   * @return new table update timestamp, or 0 if the table does not
+   * exist
+   */
+  long markDeleting(TableId id);
+
+  /**
+   * Read the table record for the given ID.
+   *
+   * @return the table record, or {@code null} if the entry is not
+   * found in the DB.
+   */
+  @Nullable TableMetadata read(TableId id);
+
+  /**
+   * Delete the table record for the given ID. Essentially does a
+   * "DELETE IF EXISTS". There is no version check. Delete should be
+   * called only when there are no segments left for the table: use
+   * {@link #markDeleting(TableId)} to indicates that the segments are
+   * being deleted. Call this method after deletion is complete.
+   * <p>
+   * Does not cascade deletes yet. Eventually, should delete all entries
+   * for the table.
+   *
+   * @return {@code true} if the table exists and was deleted,
+   * {@code false} if the table did not exist.
+   */
+  boolean delete(TableId id);
+
+  List<TableId> list();
+  List<String> list(String dbSchema);
+  List<TableMetadata> listDetails(String dbSchema);
+
+  void stop();
 }
